Tests for input_dir
WIP #1938
